requirements.txt
Django==2.2.7
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-js-asset==1.2.2
easy-thumbnails==2.6
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
